src/backend/requirements.txt
fastapi==0.111.0
pydantic==2.7.2
openai
weaviate-client